Remove redundant servlets and packages.
