+++ firebase-web/src/main/java/io/spine/web/firebase/subscription/FirebaseSubscriptionBridge.java
@@ -31,9 +31,6 @@ import io.spine.client.Topic;
 import io.spine.client.grpc.QueryServiceGrpc;
 import io.spine.web.firebase.client.FirebaseClient;
 import io.spine.web.firebase.client.NodePath;
-import io.spine.web.firebase.subscription.cancel.FirebaseSubscriptionCancelResult;
-import io.spine.web.firebase.subscription.keepup.FirebaseSubscriptionKeepUpResult;
-import io.spine.web.firebase.subscription.subscribe.FirebaseSubscribeResult;
 import io.spine.web.query.service.AsyncQueryService;
 import io.spine.web.subscription.SubscriptionBridge;
 import io.spine.web.subscription.result.SubscribeResult;
