--- firebase-web/src/main/java/io/spine/web/firebase/subscription/keepup/FirebaseSubscriptionKeepUpServlet.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.web.firebase.subscription.keepup;
-
-import io.spine.client.Subscription;
-import io.spine.web.firebase.subscription.FirebaseSubscriptionBridge;
-import io.spine.web.subscription.servlet.SubscriptionKeepUpServlet;
-
-/**
- * A {@link SubscriptionKeepUpServlet} which uses a {@link FirebaseSubscriptionBridge} to send off
- * the requests to keep up a subscription.
- *
- * @see FirebaseSubscriptionBridge#keepUp(Subscription)
- */
-@SuppressWarnings("serial") // Java serialization is not supported.
-public class FirebaseSubscriptionKeepUpServlet extends SubscriptionKeepUpServlet {
-
-    protected FirebaseSubscriptionKeepUpServlet(FirebaseSubscriptionBridge bridge) {
-        super(bridge);
-    }
-}
