+++ firebase-web/src/test/java/io/spine/web/firebase/subscription/cancel/FirebaseSubscriptionCancelResultTest.java
@@ -20,6 +20,7 @@
 
 package io.spine.web.firebase.subscription.cancel;
 
+import io.spine.web.firebase.subscription.FirebaseSubscriptionCancelResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
